requirements.txt
absl-py==0.9.0
astor==0.8.0
autocorrect==1.1.0
blinker==1.4
cachetools==3.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cycler==0.10.0
Flask==1.0.3
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
i-impute==1.0.6
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nexmo==2.4.0
nltk==3.5
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==1.0.3
protobuf==3.11.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.7
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
six==1.14.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tqdm==4.46.0
twilio==6.39.0
urllib3==1.25.8
Werkzeug==0.14.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
